+++ ignite/templates/query/files/x/{{moduleName}}/client/cli/query_{{queryName}}.go.plush
